# swagger.yml
openapi: 3.0.0
servers:
  - description: ServidorLocal
    url: 'http://localhost:3000'
info:
  version: 1.0.0
  title: Delilah Resto
  description:  API de pedidos online para un restaurante.
...
paths:
...
  /productos:
    get:
      tags:
        - Productos
      summary: Obtener lista de productos
      description: Lista todos los productos
      responses:
        200:
          description: Operacion Exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Productos"
        400:
          description: No se pudo realizar la operacion...
    post:
      tags:
        - Productos
      summary: Agregar producto (solo Admin)
      description: Agregar producto a la base de datos
      security:
        - autenticacionJWT: []
      requestBody:
        description: Crear objeto del producto
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Producto"
        required: true
      responses:
        200:
          description: Producto agregado con exito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
        400:
          description: No se pudo crear el producto
...
  /pedidos:
    get:
      tags:
        - Pedidos
      summary: Lista los pedidos
      description: Si es admin retorna todos los pedidos, si es cliente , retorna solo los pedidos del mismo
      responses:
        200:
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pedidos"
        401:
          description: No hizo ningun pedidos
        201:
          description: No hay pedidos
    post:
      tags:
        - Pedidos
      summary: Agregar pedidos
      description: Agregar pedidos a la base de datos
      requestBody:
        description: Objeto del pedido a crear
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/nuevoPedido"
        required: true
      responses:
        200:
          description: Pedido creado con exito
        400:
          description: Ha ocurrido un error
        401:
          description: Sin Stock
...
components:
  schemas:
...
    Producto:
      type: object
      properties:
        nombre:
          type: string
          example: Pizza
        descripcion:
          type: string
          example: piiza napolitana
        foto:
          type: string
          example: https://truffle-assets.imgix.net/7fae7223-pizza-napolitana_s_thumb.jpg
        stock:
          type: integer
          example: 100
        favorito:
          type: boolean
          example: false
        precio:
          type: integer
          example: 500
    Productos:
      type: array
      items:
        $ref: "#/components/schemas/Producto"
    nuevoPedido:
      type: object
      properties:
        idProducto:
          type: array
          example: 3,4,5
        cantidad:
          type: array
          example: 10,2,5
        forma_pago:
          type: string
          example: Efectivo
      Pedidos:
        type: object
        properties:
          "Nro Pedido":
            type: integer
            example: 3
          "Valor pedido":
            type: integer
            example: 1000
          detalle:
            type: string
            example: 2 Tacos
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []